requirements.txt
fastapi
uvicorn[standard]
tensorflow
python-multipart
pillow
numpy